Two small fixups
When this CCAudit message was added to mmap.c there were no braces and
they needed to be added.  Add them now.

Adding the braces around this if clause removes the need for "rc" to be
initialized.  Guessing an engineer at Qualcomm had initialized "rc" to
suppress a warning, instead of looking further and realizing there was
an actual error (fixed earlier in LineageOS).  Tiny performance increase
here.