Add candidate for H990* configuration
Since the only unlocked bootloader available for the V20 is the
well-circulated debug one, there are adjustments to make that work.  The
kernel command-line is adjusted to report the correct serial number, and
remove the "LG-US996" model name.  The model name and SIM count MUST BE
ADDED TO THE BOOT IMAGE FOR THIS KERNEL TO FUNCTION!