Add candidates for H990DS default configuration
This ends up being a pair of default configurations.  The
elsa_global_com-h990ds variant has the minimal changes to produce a near-
stock kernel which works with the unlocked bootloader.  The lineageos
variant tries to be as close to the "standard" LineageOS kernel for the
LGE V20 device.  These also aim to be like standard defconfigs and *not*
list every possible option, merely those that need to be set.

Since the only unlocked bootloader available for the V20 is the
well-circulated debug one, there are adjustments to make that work.  The
kernel command-line is fixed to report the correct model and the hardware
present (dual-SIM).  Testing has shown this has a genuine effect on
functioning of hardware features.

It was found that LGE's software *requires* the Bluetooth driver to be a
module.  Their bluetooth-manager daemon appears to hard-code an `insmod`
of the driver, and if that fails the daemon exits.  There is some
speculate of whether a module which successfully loads and then unloads
itself could be used to replace the on-storage module files, but this
has not yet been tested.

Not all settings from LGE's kernels were preserved.  Of note, LGE's
default configuration sets the following additional values:

CONFIG_LGE_BLUETOOTH_PM=y

CONFIG_SND_LGE_EFFECT=y
CONFIG_SND_LGE_NORMALIZER=y
CONFIG_SND_LGE_MABL=y

CONFIG_MSM_CORE_CTL_HELPER=y