Add H990* version of panel timings
This is the version of this file needed for H990* type devices.  This has
been confirmed to work on H990DS and H990N devices (and believed to
work on H990).  There is some speculation about only needing two versions
of this file, but volunteers to try the US Carrier version on H990*
devices weren't handy.