Fix disabling of second screen
The original change, a2ed7a25e81eeeb8ebe42d281ace647b8439204b was really
anti-social.  This is needed for LinageOS, but this breaks things for ALL
other users of this kernel source.  Turn this into a configuration
option, so it can be enabled/disabled without needing to modify the
source code.