Implementation of DirtySanta "smem" cleanup code
Due to its early debugging release nature, the "DirtySanta" bootloader
leaves some hardware portions uninitialized or incorrectly initialized.
This code works around one portion of memory which is left uninitialized.

This code relies on knowing the layout of the "smem" area.  There is some
ability to change this, but this doesn't seem too likely to be done before
a new phone model is introduced.

There are two potential sources for the values used.  The configuration
options CONFIG_DIRTYSANTA_FIXUP_DEVICENAME and
CONFIG_DIRTYSANTA_FIXUP_SIMCOUNT can be set to appropriate values.  The
second is the code can look for the values of "model.name" and
"lge.sim_num" on the kernel command-line.  In the second case these
values are likely to need to be introduced/modified elsewhere.

The handler for the "smem" area is first initialized during the
arch_initcall() phase, which places this into subsys_initcall().